Basework for Skins & Multiplatform support
GBA is bugged as of right now, but this means that we can branch out very easily to more systems in BizHawk with the creation of the skins file, a UI file and a triggers file.
A lot of the stuff (especially regarding triggers/ui is copy/paste. Most of the leg work is in the skins file. So creating new skins or support for new systems should be very simple. Aim is to get it working for NES/SNES/GBA/N64 at least)